refactor: fix authCfg data race, reduce handler boilerplate, extract logging
Fix a data race on the shared auth config and clean up five code quality
issues that clashed with the project's core principles.

Bug fix:
- Add sync.RWMutex around all authCfg reads/writes in auth_handlers.go
  and tls_handlers.go. Previously, concurrent requests could race on
  PasswordHash, Username, and TLS/ACME fields.

Consistency:
- Replace all 8 remaining json.NewDecoder(r.Body).Decode() calls with
  the project's decodeJSON() helper, which rejects trailing garbage.
- Add missing SSE publish calls after createDataset, deleteDataset,
  createSnapshot, deleteSnapshot, deleteSnapshotBatch, startScrub, and
  cancelScrub. New publishSnapshots() and publishPools() helpers.

Boilerplate:
- Add writeRunOpError() helper; replaces 33 copies of the 5-line
  error-handling block after runOp calls (~165 lines removed).
- Remove now-unused ansible import from 4 handler files.

Structure:
- Extract journalHandler and requestLogger from main.go into new
  internal/logging/ package (NewJournalHandler, RequestLogger).
  main.go drops from 345 to 180 lines.

Docs updated: CHANGELOG.md, CLAUDE.md file map, README.md and
wiki/Architecture.md diagrams.
