+++ internal/logging/middleware.go
@@ -1,0 +1,72 @@
+package logging
+
+import (
+	"crypto/rand"
+	"encoding/hex"
+	"log/slog"
+	"net/http"
+	"time"
+
+	"dumpstore/internal/api"
+)
+
+// newReqID returns a random 16-character hex string for request correlation.
+func newReqID() string {
+	var b [8]byte
+	_, _ = rand.Read(b[:])
+	return hex.EncodeToString(b[:])
+}
+
+// RequestLogger wraps a handler and emits one logfmt line per request.
+// It generates a unique req_id, stores it in the request context, and includes
+// it on the request log line so downstream slog calls can be correlated.
+func RequestLogger(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		id := r.Header.Get("X-Request-ID")
+		if id == "" {
+			id = newReqID()
+		}
+		r = r.WithContext(api.WithReqID(r.Context(), id))
+		w.Header().Set("X-Request-ID", id)
+
+		start := time.Now()
+		rw := &statusRecorder{ResponseWriter: w, status: http.StatusOK}
+		next.ServeHTTP(rw, r)
+		elapsed := time.Since(start)
+
+		level := slog.LevelInfo
+		if rw.status >= 500 {
+			level = slog.LevelError
+		} else if rw.status >= 400 {
+			level = slog.LevelWarn
+		}
+		slog.Log(r.Context(), level, "request",
+			"req_id", id,
+			"method", r.Method,
+			"path", r.URL.Path,
+			"status", rw.status,
+			"duration_ms", elapsed.Milliseconds(),
+			"remote", r.RemoteAddr,
+		)
+		api.RecordHTTP(r.Method, r.URL.Path, rw.status, elapsed)
+	})
+}
+
+// statusRecorder captures the HTTP status code written by a handler.
+type statusRecorder struct {
+	http.ResponseWriter
+	status int
+}
+
+func (r *statusRecorder) WriteHeader(code int) {
+	r.status = code
+	r.ResponseWriter.WriteHeader(code)
+}
+
+// Flush implements http.Flusher by forwarding to the underlying ResponseWriter
+// if it supports flushing. Required for SSE streaming to work through this middleware.
+func (r *statusRecorder) Flush() {
+	if f, ok := r.ResponseWriter.(http.Flusher); ok {
+		f.Flush()
+	}
+}
